+++ cmake/morpheus_utils/package_config/pybind11/Configure_pybind11.cmake
@@ -32,11 +32,12 @@ function(morpheus_utils_configure_pybind11)
     INSTALL_EXPORT_SET
       ${PROJECT_NAME}-core-exports
     CPM_ARGS
-      GIT_REPOSITORY  https://github.com/pybind/pybind11.git
-      GIT_TAG         "v${PYBIND11_VERSION}"
-      GIT_SHALLOW     TRUE
-      OPTIONS         "PYBIND11_INSTALL ON"
-                      "PYBIND11_TEST OFF"
+      GIT_REPOSITORY https://github.com/pybind/pybind11.git
+      GIT_TAG "v${PYBIND11_VERSION}"
+      GIT_SHALLOW TRUE
+      PATCH_COMMAND git checkout -- . && git apply --whitespace=fix ${CMAKE_CURRENT_FUNCTION_LIST_DIR}/patches/gil_dec_ref.patch
+      OPTIONS "PYBIND11_INSTALL ON"
+              "PYBIND11_TEST OFF"
   )
 
   list(POP_BACK CMAKE_MESSAGE_CONTEXT)